Add files via upload
I did some changes on AutoDDG code to support be run on spark when informed. My goal was to separate, we should run autoddg profiler with spark when the datasets were large, and locally with pandas when they are not, this way we could have time efficiency. On the notebook I created a list of all the datasets with their number of columns, and with that I runned the AutoDDG profiler with and without spark, for datasets with less then 10M rows, localy is better than spark, with the datasets that had more then 100M rows, we had error running locally because it is to big, and we were able to run with spark.