Add files via upload
These are some keyword and nl queries I generated using an LLM. I inputted each CSV table, and prompted the LLM:
Suppose a user wants to search for this dataset in a dataset search engine using keywords. Generate 10 keyword queries of varying granularities (less than 10 words) that a user can use to search for these datasets.
Suppose a user wants to search for this dataset in a dataset search engine using a natural language question. Generate 10 questions that a user can use to search for these datasets.

As evaluation we could evaluate if the vector search performs well for our dataset descriptions